Day8

+++ Day8.py
@@ -135,9 +135,9 @@ print("Quotient is : ", int(a) / int(b))
 # Ask the user to enter two integers and one float. Convert them all to floats
 # and print their average.
 
-a = int(input("Enter the first value: "))
-b = int(input("Enter the second value: "))
+a = float(input("Enter the first value: "))
+b = float(input("Enter the second value: "))
 c = float(input("Enter the third value: "))
 
-average = (a + b + c)/3
+average = (a + b + c) / 3
 print(average, type(average))
